Adding dbscan clustering for anomaly testing
Adding of the DBSCAN algorithm for the testing part. It makes easier to change parameters of DBSCAN directly in dbscan_prediction function (test_ensemble_anomaly.py). There is no training for DBSCAN so it can directly run on the testing dataset. 

Some other minor changes related to the new path of data directory after the merging and to the adding of powershell scripts to build / start spark.