HDBSCAN implementation (training and real-time clustering)
Implementation of the HBSCAN algorithm. Adding of the algorithm in the spark scripts, training on the train dataset and predictions from the incoming data in real-time. 

This commit also includes the notebook that I created to briefly present HDSBSCAN and to tune the parameters. The spark docker files (to build the image) have been updated to include the HDBSCAN libraries in the spark container.